Add server side fetching for admin events
