+++ src/app/(protected)/admin/events/page.tsx
@@ -3,7 +3,6 @@ import AdminEvents from "@/app/components/admin/AdminEvents";
 export default function AdminEventsPage() {
   return (
     <main>
-      {/* We simply import and render the component here */}
       <AdminEvents /> 
     </main>
   );
